requirements.txt
python-telegram-bot[all]==21.9
SQLAlchemy==2.0.36
requests==2.32.3
python-dotenv==1.0.1
python-dateutil==2.9.0
pandas==2.2.3